+++ active_adaptation/assets/underwater/BlueROVHeavy.py
@@ -102,16 +102,18 @@ def make_isaaclab_cfg(self_collisions: bool = False):
     return AssetSpec(
         config=asset_cfg,
         sensors={},
-        wrapper=UnderwaterRobot(
-            cfg=HydrodynamicsCfg(
-                volume=VOLUME,
-                coBM=COBM,
-                added_mass=ADDED_MASS,
-                linear_damping=LINEAR_DAMPING,
-                quadratic_damping=QUADRATIC_DAMPING,
+        adaptations=(
+            UnderwaterRobot(
+                cfg=HydrodynamicsCfg(
+                    volume=VOLUME,
+                    coBM=COBM,
+                    added_mass=ADDED_MASS,
+                    linear_damping=LINEAR_DAMPING,
+                    quadratic_damping=QUADRATIC_DAMPING,
+                ),
+                rotor_time_constants=ROTOR_TIME_CONSTANTS,
+                rotor_force_constants=ROTOR_FORCE_CONSTANTS,
             ),
-            rotor_time_constants=ROTOR_TIME_CONSTANTS,
-            rotor_force_constants=ROTOR_FORCE_CONSTANTS,
         ),
     )